# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket Booking System API Documentation",
        "description": "API documentation for the Ticket Booking System backend.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Ticket Booking System API Server"
        }
    ],
    "paths": {
...
        "/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all bookings for the authenticated user",
                "operationId": "b89b932c131e727a3720a01899acd0a2",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "166547aa028b063e31d6efe9c093aa88",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "from_station",
                                    "to_station",
                                    "journey_date",
                                    "seat_type",
                                    "number_of_seats",
                                    "total_fare"
                                ],
                                "properties": {
                                    "from_station": {
                                        "type": "string",
                                        "example": "Station C"
                                    },
                                    "to_station": {
                                        "type": "string",
                                        "example": "Station D"
                                    },
                                    "journey_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-30"
                                    },
                                    "seat_type": {
                                        "type": "string",
                                        "example": "Business"
                                    },
                                    "number_of_seats": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "total_fare": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get a specific booking by ID",
                "operationId": "9a7e24b15b1f9fe6c357532b3c5200a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update a specific booking by ID",
                "operationId": "620746af203837311ea0dd4fca355623",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "from_station": {
                                        "type": "string",
                                        "example": "Station E"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Booking not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Delete a specific booking by ID",
                "operationId": "b14ac6c3aa1a450e338f1424cc7d7168",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Booking deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Booking": {
                "title": "Booking",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationError"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "user_id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "from_station": {
                                "type": "string",
                                "example": "Station A"
                            },
                            "to_station": {
                                "type": "string",
                                "example": "Station B"
                            },
                            "journey_date": {
                                "type": "string",
                                "format": "date",
                                "example": "2025-12-25"
                            },
                            "seat_type": {
                                "type": "string",
                                "example": "Economy"
                            },
                            "number_of_seats": {
                                "type": "integer",
                                "example": 2
                            },
                            "total_fare": {
                                "type": "number",
                                "format": "float",
                                "example": 100
                            },
                            "status": {
                                "type": "string",
                                "example": "pending"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00.000000Z"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00.000000Z"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "BookingList": {
                "title": "Booking List",
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Booking"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationError"
                    }
                ]
            },
...
            "ValidationError": {
                "title": "Validation Error",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}